+++ RELEASE-NOTES.md
@@ -1,6 +1,6 @@
 ## 3.0.M1
 
-* #30 Kafka own Consumer as a more capbale alternative of KafkaConsumerRunner
+* #30 Replace consumer iterator with callback
 
 ## 2.0.M1